enter handler hook
