+++ src/shared/hooks/useEnterHandler.ts
@@ -1,0 +1,15 @@
+import { useEffect } from "react";
+
+export const useEnterHandler = (func: () => void) =>
+  useEffect(() => {
+    const listener = (event: any) => {
+      if (event.code === "Enter" || event.code === "NumpadEnter") {
+        event.preventDefault();
+        func();
+      }
+    };
+    document.addEventListener("keydown", listener);
+    return () => {
+      document.removeEventListener("keydown", listener);
+    };
+  }, [func]);
